+++ CHANGELOG.md
@@ -13,7 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - CategoryAxisOffset to have more control over CategoryAxisModel's positioning of categories
   relative to the origin.
 - AxisContent and GridStyle to shorten the function signature of XYGraph.
-- Support for cubic Bezier curves in line, area, and stacked area plots
+- Support for cubic Bezier curves in line, area, and stacked area plots.
 
 ### Changed
 
@@ -22,6 +22,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Deprecated Composable functions that did not comply with Compose coding standards, and implemented
   compliant replacements.
 - Renamed ConstantLine to HorizontalLine for area plot baselines.
+- XYGraph pointer input callback replaced with a pointer event handler and a scope for converting
+  pointer coordinates (pixels) to graph AxisModel coordinates.
 
 ### Fixed
 
@@ -42,7 +44,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- XYGraph pointer tracking would throw an exception if the pointer event had a position outside the bounds of the
+- XYGraph pointer tracking would throw an exception if the pointer event had a position outside the
+  bounds of the
   component.
 
 ## [0.10.2]